+++ Controllers/Schemas/ProductSchema/UpdateProduct.cs
@@ -7,7 +7,7 @@ namespace BE_Shop.Controllers
 		public Guid Id { get; set; } = Guid.Empty;
 		public string Name { get; set; } = string.Empty;
 		public string Decription { get; set; } = string.Empty;
-		public long UnitPrice { get; set; } = 0;
+		public int UnitPrice { get; set; } = 0;
 		public int TotalItem { get; set; } = 0;
 		public Guid MainFile { get; set; } = Guid.Empty;
 		public int Discount { get; set; } = 0;
